Improve Plugin: AutoClose
More filetypes with special rules for autoclosing chars.

Also, give up on trying to get the Endwise and this Enter key to be
pals, this is way above my paygrade.
The best shot is defining the AutoClose enter before EndWise, so that
the EndWise logic can work them out, but it's a longshot, and I'd have
to gut this plugin anyway.