Initial spelling configuration
This is a big commit that sets up the bare minimum spelling configuration.

The really important part is the `spell` directory, with some extra
structure to support the `plugin/spell.vim` file. This could be spun off
into it's own plugin, but for now it's still being tested. When I start
the `.add` files song and dance I'll see how robust this is.

The (supported) way to change languages is by using the `ts` mapping,
anything else is brittle. I shall create a better way to do it, by
adding to the SpellLoop plugin. Another TBA feature is integrating
the SpellCheck plugin with Syntastic, so that I don't need to hack
something involving the `Q` mapping.

The `spl` files come from the default vim repository, but I duplicate
them here so that I can update them seamlessly, using `:mkspell` or by
finding anyone that has been through that rocky path.