More Windows bugfixing
Use by default the DejaVU Sans Mono font, patched for Powerline,
because if you'll have to bother installing a new font, might as well go
the extra mile.

The folder with the fonts so that's easier to just clone this repo and
it's good to go. Doesn't add THAT much weight.