Fix AutoClose to play ball with endwise
Finally! A solution for the problem of having AutoClose and Tim Pope's
endwise both active. It's an ugly hack, since endwise does a similar
thing, but AutoClose defines things on the buffer level, it requires a
different approach.

The important thing is that it works as intended. The endwise endings
are correctly introduced, and the defined pairs on AutoClose also get a
newline in between them. Makes working on PHP almost bearable.