Introduce helpers for updating the plugins
The 3rd party script is a tried-and-true workhorse, production ready
methinks. It's easy, just fetch the last version and fast forward it.

The 1st party script is much more complicated, and I intend to keep
refining it. It hardcodes lots of stuff, it's not very flexible. It
works as long as you monitor and/or have a clear revert strategy.
It also depends on the modules file having extra meta-data. I tried
having this on the repos themselves, but submodules and extra remotes
don't like each other. Plus, it's not portable between machines.

There should be a `plugin-fork` helper, that does exactly that, but I'm
still testing it.

Another much needed helper is adding a 3rd Party plugin by specifying
the URL directly.