Sort out Airline divergence
A while back I submitted a PR for the Airline plugin that wasn't
integrated due to incompatibilities with OSX and/or the UltiSnips
plugin. I submitted a second one, cut down to the really interesting
change of defining a format and that was merged.

This commit moves my code to my private vimrc, and configures the
Airline plugin to follow the upstream, and using my version of the
feature in the vimrc.

This will avoid a maintenance nightmare of a manual merge every time
Airline updates.