Create insert mapping to save file
BETRAYAL! This is a concession to the mainstream text
editors/processors, but it's really handy. For terminal vim there should
be a similar thing, but I think I can live.
This is mostly useful in writing free text, like this very commit
message. Vim backup is not enough, I don't think it saves in between
mode changes.