Enable Python omnicompletion
It comes by default, but I hadn't exposed it in the SuperTab
configuration.
I better source/investigate a better omnicomplete function, such as:
https://github.com/mjbrownie/pythoncomplete.vim
https://github.com/vim-scripts/pythoncomplete/pull/1
but for now, this will do.

I'll have to search for a way to customize the preview window that
appears on top of the window containing the pydoc for the corresponding
function (Hot Damn! It's starting to look like a proper IDE).