Improve search settings
Make sure the visualstar plugin detection runs when the plugin has
already been defined.

Add a new mapping to create a skeleton of a command to replace the
current pattern in the whole file (or a visual selection). This is
the meat of the "vim-over" plugin, without creating an entire parallel
command line.