Configure textobj-user and plugins
This configures the text objects I installed in the last commit. As
mentioned, this obsoletes the "vim-all" plugin, which is a first for
this repository.

There's a lot of useful plugins here, and I only explored the first page
of github results, now it's just a matter of using them on a daily basis
to get them ingrained in my memory.
There's two PR I will submit:
- folds, `iz` ignoring the lines with the fold markers (DONE)
- comments, breaking up multiline comments on fold markers