New Plugin: textobj-function
This plugin formalizes what I've been trying to do, which is setting up
a single mapping to the textobj for functions in different languages. It
even defines the mapping in a single location, so it's better than my
policy-based solution.
I've removed the now redundant mappings.

This commit will also add my patch to the ruby version of the textobj function
plugin.
Also, add the same thing for javascript. That's a lot of ground covered
already.