Add C# Initial Support
This commit introduces and configures OmniSharp, to provide IDE-like
features for C#. It's my fork, I just removed the new Roslyn-based
server, I have no need for it yet, not to clutter this repo. I'll just
add it as needed.
This needs Python support, so I'm gonna disable it by default on
Windows. It's a hassle to make it work, if I really needed I'm just one
comment away.

This also adds syntax for the CSHTML views.

The only thing missing is autocompletion for CSHTML, which can leverage
the OmniSharp integration. Can't find anything that does that, that's a
nice idea for a complex plugin.