Add a new helper: Fork a plugin
This adds a very handy helper for a common operation: forking a plugin.
This involves changing urls around, moving folders and editing (and
commiting) the `.gitmodules` file.
Never again.

The next level is having backend for different cloud services. So,
forking a repo on GitHub would actually fork the repo and update the
urls. One can dream.

The `config-1st-party` is just the first pass at trying to unite the
hodgepodge that is `sh_helpers`. This scripts are not battle-hardened.