New Plugin: Insert semicolons on the right place
This adds a customizable plugin to insert semicolons in the end of the
line, just like some other IDE like NetBeans.
Currently it's only enabled for Javascript, but it's extensible enough
to work for inserting characters on correct places.

It works, but the architecture is ugly. Should properly use `ftplugin`s
and other helpers, as shift the bulk of the logic to `autoload`.
This is a candidate for spinning off to a discrete plugin.