Configure Signify plugin for non-Git VCS
The only enabled VCS for now are TFS and p4 (p4 untested).
This configuration is very conservative, check only on file write, since
checking writes the file (:S). Chalk it up to VCS limitations.

It's slow as hell on Windows, so I'll have to investigate another way to
run this: async or manual (Use <Leader>d perhaps). Probably a
buffer-local mapping for Signified buffers, but that needs plugin
cooperation. A fork is in order, add a User autocmd.