Improve the Plugins automation
This is a new script that will form the backbone of the plugins
automated updated.
It automatically setups the submodules for correct operation and updates
them on demand.
This has now feature parity with "sh_helpers/update-3rd-party", and can
be used to implement "sh_helpers/update-1st_party" with less bugs.